+++ src/api/parser.ts
@@ -1,15 +1,19 @@
-import {beginCell, Cell, Dictionary, DictionaryValue, Slice} from '@ton/core';
+import { beginCell, Cell, Dictionary, DictionaryValue, Slice } from '@ton/core';
 import {
     AssetConfig,
     AssetData,
+    ClassicOracleInfo,
     ExtendedAssetData,
     ExtendedAssetsConfig,
     ExtendedAssetsData,
     MasterConstants,
     MasterData,
     PoolAssetsConfig,
-    PoolConfig
+    PoolConfig,
+    PythOracleInfo,
 } from '../types/Master';
+import { BalanceType, UserBalance, UserData, UserLiteData, UserRewards } from '../types/User';
+import { loadMaybeMyRef, loadMyRef } from './helpers';
 import {
     bigIntMax,
     bigIntMin,
@@ -21,9 +25,6 @@ import {
     getAvailableToBorrow,
     presentValue,
 } from './math';
-import {loadMaybeMyRef, loadMyRef} from './helpers';
-import {BalanceType, UserBalance, UserData, UserLiteData, UserRewards} from '../types/User';
-import {parseFeedsMapDict} from "./feeds";
 
 export function createUserRewards(): DictionaryValue<UserRewards> {
     return {
@@ -63,7 +64,17 @@ export function createAssetData(): DictionaryValue<AssetData> {
             const trackingBorrowIndex = BigInt(src.loadUintBig(64));
             const awaitedSupply = BigInt(src.loadUintBig(64));
 
-            return { sRate, bRate, totalSupply, totalBorrow, lastAccural, balance, trackingSupplyIndex, trackingBorrowIndex, awaitedSupply };
+            return {
+                sRate,
+                bRate,
+                totalSupply,
+                totalBorrow,
+                lastAccural,
+                balance,
+                trackingSupplyIndex,
+                trackingBorrowIndex,
+                awaitedSupply,
+            };
         },
     };
 }
@@ -143,11 +154,57 @@ export function createAssetConfig(): DictionaryValue<AssetConfig> {
     };
 }
 
-export function parseMasterData(masterDataBOC: string, poolAssetsConfig: PoolAssetsConfig, masterConstants: MasterConstants): MasterData {
+export interface OracleParser {
+    parseOracleConfig(masterConfigSlice: Slice): any;
+    getIfActive(masterConfigSlice: Slice): number;
+}
+
+export class ClassicOracleParser implements OracleParser {
+    parseOracleConfig(masterConfigSlice: Slice): ClassicOracleInfo {
+        const oraclesSlice = masterConfigSlice.loadRef().beginParse();
+        return {
+            numOracles: oraclesSlice.loadUint(16),
+            threshold: oraclesSlice.loadUint(16),
+            oracles: loadMaybeMyRef(oraclesSlice),
+        };
+    }
+
+    getIfActive(masterConfigSlice: Slice): number {
+        return masterConfigSlice.loadInt(8);
+    }
+}
+
+export class PythOracleParser implements OracleParser {
+    parseOracleConfig(masterConfigSlice: Slice): PythOracleInfo {
+        const oraclesSlice = masterConfigSlice.loadRef().beginParse();
+        const feedDataCell = oraclesSlice.loadRef();
+        const feedDataSlice = feedDataCell.beginParse();
+
+        return {
+            pythAddress: oraclesSlice.loadAddress(),
+            feedsMap: feedDataSlice.loadDict(Dictionary.Keys.BigUint(256), Dictionary.Values.Buffer(64)),
+            allowedRefTokens: feedDataSlice.loadDict(Dictionary.Keys.BigUint(256), Dictionary.Values.BigUint(256)),
+            pricesTtl: oraclesSlice.loadUint(32),
+            pythComputeBaseGas: oraclesSlice.loadUintBig(64),
+            pythComputePerUpdateGas: oraclesSlice.loadUintBig(64),
+            pythSingleUpdateFee: oraclesSlice.loadUintBig(64),
+        };
+    }
+
+    getIfActive(masterConfigSlice: Slice): number {
+        return masterConfigSlice.loadInt(8);
+    }
+}
+
+export function parseMasterData(
+    masterDataBOC: string,
+    poolAssetsConfig: PoolAssetsConfig,
+    masterConstants: MasterConstants,
+    oracleParser: OracleParser,
+): MasterData {
     const masterSlice = Cell.fromBase64(masterDataBOC).beginParse();
     const meta = masterSlice.loadRef().beginParse().loadStringTail();
     const upgradeConfigParser = masterSlice.loadRef().beginParse();
-
     const upgradeConfig = {
         masterCodeVersion: Number(upgradeConfigParser.loadCoins()),
         userCodeVersion: Number(upgradeConfigParser.loadCoins()),
@@ -159,11 +216,10 @@ export function parseMasterData(masterDataBOC: string, poolAssetsConfig: PoolAss
         newUserCode: loadMaybeMyRef(upgradeConfigParser),
     };
     // upgradeConfigParser.endParse(); todo fix with new testnet contract
-
     const masterConfigSlice = masterSlice.loadRef().beginParse();
+
     const assetsConfigDict = masterConfigSlice.loadDict(Dictionary.Keys.BigUint(256), createAssetConfig());
     const assetsDataDict = masterSlice.loadDict(Dictionary.Keys.BigUint(256), createAssetData());
-
     const assetsExtendedData = Dictionary.empty<bigint, ExtendedAssetData>();
     const assetsReserves = Dictionary.empty<bigint, bigint>();
     const apy = {
@@ -171,25 +227,18 @@ export function parseMasterData(masterDataBOC: string, poolAssetsConfig: PoolAss
         borrow: Dictionary.empty<bigint, number>(),
     };
 
-    for (const [tokenSymbol, asset] of Object.entries(poolAssetsConfig)) {
+    for (const [, asset] of Object.entries(poolAssetsConfig)) {
         const assetData = calculateAssetData(assetsConfigDict, assetsDataDict, asset.assetId, masterConstants);
         assetsExtendedData.set(asset.assetId, assetData);
     }
 
-    const ifActive = masterConfigSlice.loadInt(8);
-    const oraclesSlice = masterConfigSlice.loadRef().beginParse();
+    const ifActive = oracleParser.getIfActive(masterConfigSlice);
+    const oraclesInfo = oracleParser.parseOracleConfig(masterConfigSlice);
 
     const masterConfig = {
-        ifActive: ifActive,
-        oraclesInfo:  {
-            pythAddress: oraclesSlice.loadAddress(),
-            feedsMap: oraclesSlice.loadDict(Dictionary.Keys.BigUint(256), Dictionary.Values.Buffer(64)),
-            pricesTtl: oraclesSlice.loadUint(32),
-            pythComputeBaseGas: oraclesSlice.loadUintBig(64),
-            pythComputePerUpdateGas: oraclesSlice.loadUintBig(64),
-            pythSingleUpdateFee: oraclesSlice.loadUintBig(64),
-        },
+        ifActive,
         admin: masterConfigSlice.loadAddress(),
+        oraclesInfo,
         tokenKeys: loadMaybeMyRef(masterConfigSlice),
         supervisor: masterConfigSlice.loadMaybeAddress(),
     };
@@ -200,11 +249,9 @@ export function parseMasterData(masterDataBOC: string, poolAssetsConfig: PoolAss
         const totalSupply = calculatePresentValue(assetData.sRate, assetData.totalSupply, masterConstants);
         const totalBorrow = calculatePresentValue(assetData.bRate, assetData.totalBorrow, masterConstants);
         assetsReserves.set(asset.assetId, assetData.balance - totalSupply + totalBorrow);
-
         apy.supply.set(asset.assetId, (1 + (Number(assetData.supplyInterest) / 1e12) * 24 * 3600) ** 365 - 1);
         apy.borrow.set(asset.assetId, (1 + (Number(assetData.borrowInterest) / 1e12) * 24 * 3600) ** 365 - 1);
     }
-
     return {
         meta: meta,
         upgradeConfig: upgradeConfig,
@@ -221,7 +268,7 @@ export function parseUserLiteData(
     assetsData: ExtendedAssetsData,
     assetsConfig: ExtendedAssetsConfig,
     poolConfig: PoolConfig,
-    applyDust: boolean = false
+    applyDust: boolean = false,
 ): UserLiteData {
     const poolAssetsConfig = poolConfig.poolAssetsConfig;
     const masterConstants = poolConfig.masterConstants;
@@ -264,7 +311,7 @@ export function parseUserLiteData(
         let principal = realPrincipals.get(asset.assetId) || 0n;
         let balance = presentValue(assetData.sRate, assetData.bRate, principal, masterConstants);
 
-        if (applyDust && (principal > 0 && (principal < assetConfig.dust))) {
+        if (applyDust && principal > 0 && principal < assetConfig.dust) {
             principal = 0n;
             balance = {
                 amount: 0n,
@@ -304,7 +351,7 @@ export function parseUserData(
     assetsConfig: ExtendedAssetsConfig,
     prices: Dictionary<bigint, bigint>,
     poolConfig: PoolConfig,
-    applyDust: boolean = false
+    applyDust: boolean = false,
 ): UserData {
     userLiteData.fullyParsed = true;
     let havePrincipalWithoutPrice = false;
@@ -335,7 +382,7 @@ export function parseUserData(
         let principal = userLiteData.principals.get(asset.assetId) || 0n;
         const balance = presentValue(assetData.sRate, assetData.bRate, principal, masterConstants);
 
-        if (applyDust && (principal > 0 && (principal < assetConfig.dust))) {
+        if (applyDust && principal > 0 && principal < assetConfig.dust) {
             principal = 0n;
             userLiteData.principals.set(asset.assetId, 0n);
         }
@@ -359,7 +406,13 @@ export function parseUserData(
         }
     }
 
-    const availableToBorrow = getAvailableToBorrow(assetsConfig, assetsData, userLiteData.realPrincipals, prices, masterConstants);
+    const availableToBorrow = getAvailableToBorrow(
+        assetsConfig,
+        assetsData,
+        userLiteData.realPrincipals,
+        prices,
+        masterConstants,
+    );
 
     for (const [_, asset] of Object.entries(poolAssetsConfig)) {
         const balance = userLiteData.balances.get(asset.assetId) as UserBalance;
@@ -371,7 +424,17 @@ export function parseUserData(
         if (balance.type === BalanceType.supply) {
             withdrawalLimits.set(
                 asset.assetId,
-                bigIntMin(calculateMaximumWithdrawAmount(assetsConfig, assetsData, userLiteData.realPrincipals, prices, masterConstants, asset.assetId), assetData.balance)
+                bigIntMin(
+                    calculateMaximumWithdrawAmount(
+                        assetsConfig,
+                        assetsData,
+                        userLiteData.realPrincipals,
+                        prices,
+                        masterConstants,
+                        asset.assetId,
+                    ),
+                    assetData.balance,
+                ),
             );
         }
 
@@ -382,7 +445,13 @@ export function parseUserData(
 
         borrowLimits.set(
             asset.assetId,
-            bigIntMax(0n, bigIntMin((availableToBorrow * 10n ** assetConfig.decimals) / prices.get(asset.assetId)!, assetLiquidityMinusReserves)),
+            bigIntMax(
+                0n,
+                bigIntMin(
+                    (availableToBorrow * 10n ** assetConfig.decimals) / prices.get(asset.assetId)!,
+                    assetLiquidityMinusReserves,
+                ),
+            ),
         );
     }
 
@@ -395,7 +464,13 @@ export function parseUserData(
     let healthFactor = 1;
     let liquidationData;
     if (!havePrincipalWithoutPrice) {
-        liquidationData = calculateLiquidationData(assetsConfig, assetsData, userLiteData.realPrincipals, prices, poolConfig);
+        liquidationData = calculateLiquidationData(
+            assetsConfig,
+            assetsData,
+            userLiteData.realPrincipals,
+            prices,
+            poolConfig,
+        );
         if (liquidationData.totalLimit != 0n) {
             healthFactor = 1 - Number(liquidationData.totalDebt) / Number(liquidationData.totalLimit);
         }
@@ -411,6 +486,6 @@ export function parseUserData(
         limitUsed: limitUsed,
         liquidationData: liquidationData,
         healthFactor: healthFactor,
-        havePrincipalWithoutPrice: havePrincipalWithoutPrice
+        havePrincipalWithoutPrice: havePrincipalWithoutPrice,
     };
 }