+++ src/api/prices.ts
@@ -1,31 +1,8 @@
-import {PoolConfig} from '../types/Master';
-import {MAINNET_POOL_CONFIG} from '../constants/pools';
-
-import {Buffer} from "buffer";
-import {HermesClient, HexString} from "@pythnetwork/hermes-client";
-import {Cell} from "@ton/core";
-import {beginCell} from "@ton/ton";
-import {createCellChain} from "@pythnetwork/pyth-ton-js";
-import { DefaultPriceSourcesConfig, PriceSourcesConfig } from '../prices';
-
-
-export const DEFAULT_HERMES_ENDPOINT = 'https://hermes.pyth.network';
-
-/**
- * Updates feeds data from specified endpoint
- * @param hermesEndpoint prices servide endpoint
- * @param feedIds list of pyth feed ids to fetch
- * @returns binary - buffer of feeds update, parsed - json feeds data
- */
-export async function getPythFeedsUpdates(feedIds: HexString[], hermesEndpoint: string) {
-    const hermesClient = new HermesClient(hermesEndpoint);
-    const latestPriceUpdates = await hermesClient.getLatestPriceUpdates(feedIds, {encoding: 'hex'});
-
-    const parsed = latestPriceUpdates.parsed;
-    const binary = Buffer.from(latestPriceUpdates.binary.data[0], 'hex');
-
-    return {binary, parsed};
-}
+import { HexString } from '@pythnetwork/hermes-client';
+import { createCellChain } from '@pythnetwork/pyth-ton-js';
+import { Cell } from '@ton/core';
+import { beginCell } from '@ton/ton';
+import { Buffer } from 'buffer';
 
 export function composeFeedsCell(feeds: HexString[]): Cell {
     if (feeds.length === 0) {
@@ -33,13 +10,11 @@ export function composeFeedsCell(feeds: HexString[]): Cell {
     }
 
     const reversedTail = feeds.slice(1).reverse();
-    const packedTail = reversedTail.reduce(
-        (prev: Cell | null, curr) => {
-            const builder = beginCell().storeUint(BigInt(curr), 256);
-            if (prev !== null) builder.storeRef(prev);
-            return builder.endCell();
-        }, null
-    );
+    const packedTail = reversedTail.reduce((prev: Cell | null, curr) => {
+        const builder = beginCell().storeUint(BigInt(curr), 256);
+        if (prev !== null) builder.storeRef(prev);
+        return builder.endCell();
+    }, null);
     const firstFeed = feeds[0];
     const builder = beginCell().storeUint(feeds.length, 8).storeUint(BigInt(firstFeed), 256);
     if (packedTail !== null) {
@@ -49,21 +24,8 @@ export function composeFeedsCell(feeds: HexString[]): Cell {
     return builder.endCell();
 }
 
-export function packPythUpdatesData(pythUpdates: Buffer|Cell): Cell {
-    return  pythUpdates instanceof Cell ? pythUpdates : createCellChain(pythUpdates);
-}
-
-// todo: implement
-
-// need to have a mapping evaa => [pyth_original, pyth_referred] and combine all feeds to a resulting list
-/**
- * takes evaa ids and returns required set of pyth feeds
- * @param evaaIds list of evaa ids
- * @returns list of pyth feeds required to get specified evaa ids
- */
-export function createRequiredFeedsList(evaaIds: bigint[]): HexString[] {
-    throw new Error('Not implemented yet');
-    return [];
+export function packPythUpdatesData(pythUpdates: Buffer | Cell): Cell {
+    return pythUpdates instanceof Cell ? pythUpdates : createCellChain(pythUpdates);
 }
 
 /*
@@ -83,4 +45,4 @@ export async function getPrices(endpoints: string[] = ["api.stardust-mainnet.iot
 
     return { dict: prices.dict, dataCell: prices.dataCell };
 }
-*/ 
+*/