+++ src/constants/assets/index.ts
@@ -1,0 +1,3 @@
+export * from './assetId';
+export * from './mainnet';
+export * from './testnet';